Handle deleted files and directories when creating distribution
The current implementation of processFile() does not handled "unlink"
and "unlinkDir" filesystem events. When a file or directory is deleted,
and exception is thrown.

To recreate the problem:

- `touch app/assets/js/foo.js`
- `rm app/assets/js/foo.js`

Two bad things happen:

- An exception is thrown
- The file is never deleted from `_remake/dist/`

This is especially problematic for CSS and JS files that being with `_`
because the bundle is not recreated when files like that are deleted.